+++ checkconfig.php
@@ -14,7 +14,7 @@ use SebLucas\Cops\Calibre\Database;
 use SebLucas\Cops\Input\Request;
 use SebLucas\Cops\Input\Route;
 
-require_once __DIR__ . '/config.php';
+require_once __DIR__ . '/config/config.php';
 /** @var array<mixed> $config */
 
 $request = new Request();
@@ -260,7 +260,7 @@ foreach (Database::getDbList() as $name => $database) {
         echo $name . ' File ' . Database::getDbFileName($i) . ' not found,
 Please check
 <ul>
-<li>Value of $config[\'calibre_directory\'] in config_local.php <strong>(Does it end with a \'/\'?)</strong></li>
+<li>Value of $config[\'calibre_directory\'] in config/local.php <strong>(Does it end with a \'/\'?)</strong></li>
 <li>Value of <a href="http://php.net/manual/en/ini.core.php#ini.open-basedir">open_basedir</a> in your php.ini</li>
 <li>The access rights of the Calibre Database</li>
 <li>Synology users please read <a href="https://github.com/seblucas/cops/wiki/Howto---Synology">this</a></li>