+++ src/Input/Route.php
@@ -31,7 +31,7 @@ class Route
     /** @var array<string, mixed> */
     protected static $routes = [];
     /** @var string[] */
-    protected static $skipPrefix = ['index', 'json', 'fetch', 'restapi', 'graphql'];  // @todo handle 'json' routes correctly - see util.js
+    protected static $skipPrefix = ['index', 'json', 'fetch', 'restapi', 'graphql', 'phpunit'];  // @todo handle 'json' routes correctly - see util.js
     /** @var Dispatcher|null */
     protected static $dispatcher = null;
     /** @var array<string, mixed> */
@@ -241,7 +241,7 @@ class Route
     {
         $handler ??= 'index';
         // @todo take into account handler when building page url, e.g. feed or zipper
-        if (Config::get('use_route_urls') && !in_array($handler, ['index', 'json', 'restapi', 'graphql'])) {
+        if (Config::get('use_route_urls') && !in_array($handler, ['index', 'json', 'restapi', 'graphql', 'phpunit'])) {
             $params[Route::HANDLER_PARAM] = $handler;
         } else {
             unset($params[Route::HANDLER_PARAM]);
@@ -616,7 +616,7 @@ class Route
     }
 
     /**
-     * Check if we have trusted proxies defined in config_local.php
+     * Check if we have trusted proxies defined in config/local.php
      * @see https://github.com/symfony/symfony/blob/7.1/src/Symfony/Component/HttpKernel/Kernel.php#L741
      * @return bool
      */