+++ tests/GraphQLHandlerTest.php
@@ -11,7 +11,7 @@ namespace SebLucas\Cops\Tests;
 
 use SebLucas\Cops\Handlers\GraphQLHandler;
 
-require_once __DIR__ . '/config_test.php';
+require_once dirname(__DIR__) . '/config/test.php';
 use PHPUnit\Framework\TestCase;
 use SebLucas\Cops\Calibre\Database;
 use SebLucas\Cops\Framework;