+++ tests/WebDriverTest.php
@@ -29,7 +29,7 @@
 
 namespace SebLucas\Cops\Tests;
 
-require_once __DIR__ . '/config_test.php';
+require_once dirname(__DIR__) . '/config/test.php';
 //use PHPUnit\Framework\TestCase;
 use Facebook\WebDriver\Remote\RemoteWebDriver;
 use Facebook\WebDriver\WebDriverBy;