+++ tests/ZipperTest.php
@@ -10,7 +10,7 @@ namespace SebLucas\Cops\Tests;
 
 use SebLucas\Cops\Output\Zipper;
 
-require_once __DIR__ . '/config_test.php';
+require_once dirname(__DIR__) . '/config/test.php';
 use PHPUnit\Framework\TestCase;
 use SebLucas\Cops\Calibre\Book;
 use SebLucas\Cops\Calibre\Database;